requirements.txt
streamlit
requests
langgraph
python-dotenv
pymongo
pandas
openai>=1.30.0